requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
PyYAML==6.0.1
torch==2.2.2
torchvision==0.17.2
tqdm>=4.41
scipy>=1.13
wandb>=0.17.2
numpy<2